+++ ChangeLog.rst
@@ -7,7 +7,7 @@ Note worthy changes
 - If ``ACCOUNT_LOGIN_ATTEMPTS_LIMIT`` is set and the user successfully
   resets their password, the timeout is cleared to allow immediate login.
 
-- New provider: Zoom, Stocktwits
+- New provider: Zoom, Stocktwits, Zoho
 
 0.42.0 (2020-05-24)
 *******************